Initial commit of connectorToPointCloudSource
This commit adds the first version of the connectorToPointCloudSource
Glyph script which converts selected connectors into point cloud
source(s). I have tested this a little and it seems to work so far.

Each node in the connector(s) is added as a point in the point cloud
source and is assigned a spacing which is calculated by averaging the
edge lengths of each edge connected to the node. Currently, the decay
rate is hardcoded to 0.95.